requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
xlrd==2.0.1
xlwt==1.3.0
